renamed show_animation_dialog
